+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [v1.1.1]
+### Added
+- Nextclade typing for both reference-based and optional _de novo_ consensus assemblies.
+- Clade I reference sequence (`NC_003310.1`) is now available for selection.
+### Changed
+- Default reference sequence changed to `NC_003310.1` (Clade I) following mpox PHEIC announcement.
+
 ## [v1.1.0]
 ### Added
 - Support for uBAM (unaligned BAM) input files.